// src/generated/model/sivilstatusFrontend.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type {SivilstatusFrontendSivilstatus} from "./sivilstatusFrontendSivilstatus";
import type {EktefelleFrontend} from "./ektefelleFrontend";

export interface SivilstatusFrontend {
    kildeErSystem?: boolean;
    sivilstatus?: SivilstatusFrontendSivilstatus;
    ektefelle?: EktefelleFrontend;
    harDiskresjonskode?: boolean;
    borSammenMed?: boolean;
    erFolkeregistrertSammen?: boolean;
}